+++ tests/Config/ConfigBuilderTest.php
@@ -38,6 +38,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('never', $config->getCommitAuto());
         static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
 
         static::assertCount(0, $this->SUT->getWarnings());
     }
@@ -53,6 +54,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('never', $config->getCommitAuto());
         static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
 
         static::assertCount(1, $this->SUT->getWarnings());
         static::assertContains('Invalid value "foo" for option "commit-auto"', $this->SUT->getWarnings()[0]);
@@ -70,6 +72,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('never', $config->getCommitAuto());
         static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
 
         static::assertCount(1, $this->SUT->getWarnings());
         static::assertContains('"commit-bin-file" is specified but "commit-auto" option is set to "never". Ignoring.', $this->SUT->getWarnings()[0]);
@@ -87,6 +90,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('always', $config->getCommitAuto());
         static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
 
         static::assertCount(1, $this->SUT->getWarnings());
         static::assertContains('The file pointed by the option "commit-bin-file" was not found. Ignoring.', $this->SUT->getWarnings()[0]);
@@ -103,9 +107,27 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('ask', $config->getCommitAuto());
         static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
 
         static::assertCount(1, $this->SUT->getWarnings());
         static::assertContains('"commit-auto" is set to "ask" but "commit-bin-file" was not specified.', $this->SUT->getWarnings()[0]);
+    }
+
+    public function test_it_warns_when_gitlab_hosts_is_not_an_array()
+    {
+        $extra = [
+            'gitlab-hosts' => 'gitlab.company1.com',
+        ];
+
+        $config = $this->SUT->build($extra, __DIR__);
+
+        static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
+        static::assertSame('never', $config->getCommitAuto());
+        static::assertNull($config->getCommitBinFile());
+        static::assertEmpty($config->getGitlabHosts());
+
+        static::assertCount(1, $this->SUT->getWarnings());
+        static::assertContains('"gitlab-hosts" is specified but should be an array. Ignoring.', $this->SUT->getWarnings()[0]);
     }
 
     public function test_it_accepts_valid_setup()
@@ -113,6 +135,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         $extra = [
             'commit-auto' => 'ask',
             'commit-bin-file' => self::COMMIT_BIN_FILE,
+            'gitlab-hosts' => ['gitlab.company1.com', 'gitlab.company2.com'],
         ];
 
         $config = $this->SUT->build($extra, __DIR__);
@@ -120,6 +143,7 @@ class ConfigBuilderTest extends \PHPUnit_Framework_TestCase
         static::assertInstanceOf('Pyrech\ComposerChangelogs\Config\Config', $config);
         static::assertSame('ask', $config->getCommitAuto());
         static::assertSame($this->absoluteCommitBinFile, $config->getCommitBinFile());
+        static::assertCount(2, $config->getGitlabHosts());
 
         static::assertCount(0, $this->SUT->getWarnings());
     }